+++ JPEG.NT/Compressors/Compressor.cs
@@ -12,8 +12,8 @@ namespace JPEG.NT.Compressors
 {
     public class Compressor : ICompressor
     {
-        private readonly IFreqTransformer<double> freqTransformer;
-        private readonly IQuantanizer<double, byte> quantanizer;
+        private readonly IFreqTransformer<float> freqTransformer;
+        private readonly IQuantanizer<float, byte> quantanizer;
         private readonly IPacker<byte> packer;
 
         public int CompressQuality
@@ -25,8 +25,8 @@ namespace JPEG.NT.Compressors
         public int DCTSize { get; set; }
 
 
-        public Compressor(IFreqTransformer<double> freqTransformer,
-            IQuantanizer<double, byte> quantanizer,
+        public Compressor(IFreqTransformer<float> freqTransformer,
+            IQuantanizer<float, byte> quantanizer,
             IPacker<byte> packer)
         {
             DCTSize = 8;
@@ -45,7 +45,7 @@ namespace JPEG.NT.Compressors
             for (var i = 0; i < matrix.Width; i += DCTSize)
             {
                 var subMatrix = matrix.GetSubMatrix(j, DCTSize, i, DCTSize);
-                var componentSelector = new Func<Pixel, double>[] {p => p.Y, p => p.Cb, p => p.Cr};
+                var componentSelector = new Func<Pixel, float>[] {p => p.Y, p => p.Cb, p => p.Cr};
                 foreach (var channel in subMatrix.GetColorChannels(componentSelector,-128))
                 {
                     var channelFreqs = freqTransformer.FreqTransform2D(channel);
